requirements.txt
fastapi
uvicorn
numpy
Pillow
tensorflow
gdown
python-multipart
